// src/models/enums/enums.ts
export const enum RedisDBType {
  SocketClient = 'UserNotification',
}

export const enum StageType {
  Production = 'production',
  Development = 'development',
  Testing = 'test',
}

export enum LoggerLevels {
  Debug = 'debug',
  Info = 'info',
  Warn = 'warn',
  Error = 'error',
  Fatal = 'fatal',
}